LayoutTests/fast/js/pic/cached-getter-setter.html

<p>
This page tests cached access to getters and setters. If the test passes,
you'll see a series of PASS messages below.
</p>

<pre id="console"></pre>

<script>
(function() {
    if (window.layoutTestController)
        layoutTestController.dumpAsText();

    function log(s)
    {
        if (this.document)
            document.getElementById("console").appendChild(document.createTextNode(s + "\n"));
        else
            print(s + "\n");
    }
    
    function shouldBe(a, aDescription, b)
    {
        if (a === b) {
            log("PASS: " + aDescription + " should be " + b + " and is.");
        } else {
            log("FAIL: " + aDescription + " should be " + b + " but instead is " + a + ".");
        }
    }
    
    (function() {
        var o = {
            x_ : 0,
            y_ : 0,

            get x() { return this.x_; },
            set x(x) { this.x_ = x; },

            __proto__ : {
                get y() { return this.y_; },
                set y(y) { this.y_ = y; }
            }
        };

        for (var i = 0; i < 3; ++i) {
            o.x = i;
            shouldBe(o.x, "o.x", i);
        }

        for (var i = 0; i < 3; ++i) {
            o.y = i;
            shouldBe(o.y, "o.y", i);
        }
    })();

    (function() {
        var o = {
            x : 1,
            y : 1
        };
        
        function getX(o)
        {
            return o.x;
        }
        
        function setX(o, x)
        {
            o.x = x;
        }
        
        getX(o);
        getX(o);
        
        setX(o, 2);
        setX(o, 2);

        o.__defineGetter__("x", function() { return 3; });
        o.__defineSetter__("x", function(x) { });
        
        shouldBe(getX(o), "getX(o)", 3);
        setX(o, 2);
        shouldBe(getX(o), "getX(o)", 3);
    })();
})();
</script>
