LayoutTests/fast/js/pic/cached-getter-setter-expected.txt

This page tests cached access to getters and setters. If the test passes, you'll see a series of PASS messages below.

PASS: o.x should be 0 and is.
PASS: o.x should be 1 and is.
PASS: o.x should be 2 and is.
PASS: o.y should be 0 and is.
PASS: o.y should be 1 and is.
PASS: o.y should be 2 and is.
PASS: getX(o) should be 3 and is.
PASS: getX(o) should be 3 and is.
